# openapi.yaml
openapi: 3.0.3
info:
  title: tt-data-league-api
  version: v1
  description: |
    OpenAPI 3 specification inferred from the `tt-data-league-api-rest` module.
    This file was generated from controller and DTO records and includes
    example payloads and required properties where applicable.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server (inferred)
...
paths:
...
  /season_player/{id}:
    get:
      tags: [Season Player API]
      summary: Get season player by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Season player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonPlayerDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Season Player API]
      summary: Delete season player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted (no content)
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /season_player/search_by_name/{username}:
    get:
      tags: [Season Player API]
      summary: Search season players by similar name
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of season players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonPlayerDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /season_player/search_by_names/{names}:
    get:
      tags: [Season Player API]
      summary: Search season players by multiple names (comma-separated)
      parameters:
        - name: names
          in: path
          required: true
          schema:
            type: string
          description: Comma-separated list of names (e.g. "Alice,Bob")
      responses:
        '200':
          description: List of season players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonPlayerDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /season_player/find_by_license:
    post:
      tags: [Season Player API]
      summary: Find season players by license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseDto'
            example:
              licenseId: "L-1234"
              licenseTag: "senior"
      responses:
        '200':
          description: List of season players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonPlayerDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /season_player/find_by_practicioner/{practicionerId}:
    get:
      tags: [Season Player API]
      summary: Find season players by practicioner id
      parameters:
        - name: practicionerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of season players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonPlayerDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/register:
    post:
      tags: [Auth API]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '200':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth API]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth API]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    SeasonPlayerDto:
      type: object
      required: [clubMemberId, licenseId, licenseTag, yearRange]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000050"
        clubMemberId:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000002"
        licenseId:
          type: string
          example: "L-1234"
        licenseTag:
          type: string
          example: "senior"
        yearRange:
          type: string
          example: "2023-2024"
    LicenseDto:
      type: object
      required: [licenseId]
      properties:
        licenseId:
          type: string
          example: "L-1234"
        licenseTag:
          type: string
          example: "senior"
...
    UserRegistrationRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "user123"
        password:
          type: string
          example: "password123"
    UserLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "user123"
        password:
          type: string
          example: "password123"
    UserDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "user123"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        type:
          type: string
          example: "Bearer"
        username:
          type: string
          example: "user123"
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "INTERNAL_ERROR"
        message:
          type: string
          example: "An unexpected server error occurred"
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []